Let’s rework the swagger endpoint.
I think we should add a GitHub action that can run and generate the schema. If the generated schema is different from the committed change, then it can fail and add a comment. We can then have a second action that can be triggered by adding a comment, perhaps “generate openapi spec” or something (case agnostic) that will run the generation and commit it. Of course we should also make it possible to manually run this command and generate the spec changes.

To make this work we might need the spec a json or xaml file to make it easy to compare.